refactor: reorganise frontend into feature-based folder structure
