+++ toddy_shop_frontend/src/features/explore/data.ts
