+++ toddy_shop_frontend/src/features/home/components/HowItWorks.tsx
