+++ toddy_shop_frontend/src/features/shops/components/ShopCard.tsx
